Add files via upload
The code had the ability to read in the
digital words supplied by the mcc1208LS board to the shaft encoder of the water wheel. Import packages were used to be able to read
the information for the board. This information was a record log of the water wheel. The code
achieved this by a loop command, for an input number of samples. The board read in the high byte
from the quadrature decoder and stored this value. The board then repeated this step for the low
byte. The code then combined the two recordings into a single byte, which it then would convert
into a value for the angle, θ. To find the value for the angular velocity, ω , the code would compare
the current value for θ against the previous value stored for the angle. When the code finishes
running, it saves the values for θ and ω compared with time as a .txt file.